[libgens] sound/Ym2612.cpp: Use uint8_t for internal registers.
I was using int before because the registers were being initialized to -1,
but now I realized that this was basically a cheap way to initialize all
the registers to 0xFF, so we only need uint8_t instead of int.

Optimized register initialization using memset() instead of a for loop.

+++ src/libgens/sound/Ym2612.hpp
@@ -244,7 +244,7 @@ class Ym2612
 			unsigned int Inter_Step;	// Interpolation Step
 			Ym2612_Channel CHANNEL[6];	// Les 6 voies du YM2612
 			
-			int REG[2][0x100];	// Sauvegardes des valeurs de tout les registres, c'est facultatif
+			uint8_t REG[2][0x100];	// Sauvegardes des valeurs de tout les registres, c'est facultatif
 						// cela nous rend le débuggage plus facile
 		};
